fix: kuwo use uint32_t for its resource id

+++ include/parakeet-crypto/transformer/kuwo.h
@@ -13,6 +13,6 @@ constexpr size_t kKuwoDecryptionKeySize = 0x20;
 std::unique_ptr<ITransformer> CreateKuwoDecryptionTransformer(const uint8_t *key);
 std::unique_ptr<ITransformer> CreateKuwoDecryptionTransformer(const uint8_t *key, std::vector<uint8_t> v2_key);
 
-std::unique_ptr<ITransformer> CreateKuwoEncryptionTransformer(const uint8_t *key, uint64_t resource_id);
+std::unique_ptr<ITransformer> CreateKuwoEncryptionTransformer(const uint8_t *key, uint32_t resource_id);
 
 } // namespace parakeet_crypto::transformer
